Flask App/requirements.txt
boto3==1.26.111
botocore==1.29.111
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
fsspec==2023.4.0
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
mysqlclient==2.1.1
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
s3fs==0.4.2
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
